refactor essay creation and add summaries to service and repository methods
