+++ WriterApi/src/Essays.Writer.Application/Services/EssayWriterService.cs
@@ -4,16 +4,28 @@ using Essays.Writer.Application.Services.Interfaces;
 
 namespace Essays.Writer.Application.Services;
 
-public class EssayWriterService(IEssayCacheService essayCacheService, IEssayWriterRepository essayWriterRepository) : IEssayWriterService
+public class EssayWriterService(IEssayCacheService essayCacheService, IEssayWriterRepository essayWriterRepository, IAuthorRepository authorRepository) : IEssayWriterService
 {
+    /// <inheritdoc/>
     public async Task<Essay?> CreateEssay(Essay essay, CancellationToken cancellationToken = default)
     {
+        var processedAuthors = await authorRepository.EnsureAuthors(essay.Authors, cancellationToken);
+
+        essay.Authors.Clear();
+        essay.Authors.AddRange(processedAuthors);
+
         var createdEssay = await essayWriterRepository.CreateEssay(essay, cancellationToken);
         return createdEssay;
     }
 
+    /// <inheritdoc/>
     public async Task<Essay?> UpdateEssay(Essay essay, CancellationToken cancellationToken = default)
     {
+        var processedAuthors = await authorRepository.EnsureAuthors(essay.Authors, cancellationToken);
+
+        essay.Authors.Clear();
+        essay.Authors.AddRange(processedAuthors);
+
         var updatedEssay = await essayWriterRepository.UpdateEssay(essay, cancellationToken);
 
         if (updatedEssay is not null)
@@ -24,6 +36,7 @@ public class EssayWriterService(IEssayCacheService essayCacheService, IEssayWrit
         return updatedEssay;
     }
 
+    /// <inheritdoc/>
     public async Task<bool> DeleteEssay(Guid id, CancellationToken cancellationToken = default)
     {
         var isDeleted = await essayWriterRepository.DeleteEssay(id, cancellationToken);
